# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger definition for MedRoom backend.",
        "title": "MedRoom API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "Martín Saavedra",
            "url": "https://www.swagger.io/support",
            "email": "martin.saavedra@mail.udp.cl"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/puntajes": {
            "get": {
                "description": "Lista todos los puntajes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Puntajes"
                ],
                "summary": "Lista de puntajes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SwaggerMessages.ListPuntajesSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Genera un nuevo puntaje con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Puntajes"
                ],
                "summary": "Agrega un nuevo puntaje",
                "parameters": [
                    {
                        "description": "Puntaje a agregar",
                        "name": "input_puntaje",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.AddNewPuntajePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.AddNewPuntajeSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
        "/puntajes/{id_puntaje}": {
            "get": {
                "description": "Obtiene un puntaje según su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Puntajes"
                ],
                "summary": "Obtiene un puntaje",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del puntaje a buscar",
                        "name": "id_puntaje",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.GetOnePuntajeSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Modifica un puntaje con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Puntajes"
                ],
                "summary": "Modifica un puntaje",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del puntaje a modificar",
                        "name": "id_puntaje",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Puntaje a modificar",
                        "name": "input_actualiza_puntaje",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.PutOnePuntajePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.PutOnePuntajeSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un puntaje con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Puntajes"
                ],
                "summary": "Elimina un puntaje",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del puntaje a eliminar",
                        "name": "id_puntaje",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.DeletePuntajeSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ApiHelpers.ResponseError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.AddNewPuntajePayload": {
            "type": "object",
            "properties": {
                "calificacion_puntaje": {
                    "type": "integer"
                },
                "feedback_puntaje": {
                    "type": "string"
                },
                "id_evaluacion": {
                    "type": "integer"
                },
                "nombre_competencia_puntaje": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.PutOnePuntajePayload": {
            "type": "object",
            "properties": {
                "calificacion_puntaje": {
                    "type": "integer"
                },
                "feedback_puntaje": {
                    "type": "string"
                },
                "id_evaluacion": {
                    "type": "integer"
                },
                "nombre_competencia_puntaje": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.AddNewPuntajeResponse": {
            "type": "object",
            "properties": {
                "calificacion_puntaje": {
                    "type": "integer"
                },
                "feedback_puntaje": {
                    "type": "string"
                },
                "id_evaluacion": {
                    "type": "integer"
                },
                "nombre_competencia_puntaje": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.DeletePuntajeResponse": {
            "type": "object",
            "properties": {
                "calificacion_puntaje": {
                    "type": "integer"
                },
                "feedback_puntaje": {
                    "type": "string"
                },
                "id_evaluacion": {
                    "type": "integer"
                },
                "nombre_competencia_puntaje": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetOnePuntajeResponse": {
            "type": "object",
            "properties": {
                "calificacion_puntaje": {
                    "type": "integer"
                },
                "feedback_puntaje": {
                    "type": "string"
                },
                "id_evaluacion": {
                    "type": "integer"
                },
                "nombre_competencia_puntaje": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.ListPuntajesResponse": {
            "type": "object",
            "properties": {
                "calificacion_puntaje": {
                    "type": "integer"
                },
                "feedback_puntaje": {
                    "type": "string"
                },
                "id_evaluacion": {
                    "type": "integer"
                },
                "nombre_competencia_puntaje": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.PutOnePuntajeResponse": {
            "type": "object",
            "properties": {
                "calificacion_puntaje": {
                    "type": "integer"
                },
                "feedback_puntaje": {
                    "type": "string"
                },
                "id_evaluacion": {
                    "type": "integer"
                },
                "nombre_competencia_puntaje": {
                    "type": "string"
                }
            }
        },
...
        "SwaggerMessages.AddNewPuntajeSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.AddNewPuntajeResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.DeletePuntajeSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.DeletePuntajeResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.GetOnePuntajeSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.GetOnePuntajeResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.ListPuntajesSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.ListPuntajesResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.PutOnePuntajeSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.PutOnePuntajeResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}